Move DGroups to be a GroupManager
